Unit/month09

+++ src/main/java/com/ubic/shop/dto/ProductResponseDto.java
@@ -9,7 +9,7 @@ public class ProductResponseDto {
     private Long id;
     private String name;
     private int price;
-    private int stockQuantity;
+    private long stockQuantity;
     private Long categoryId;
     private String description;
     private String imgUrl;
